refactors file structure so that controllers and routes are separated
